Tidy up Extract Method fourslash tests
Partial port of ea4fdc5a9
