+++ tests/baselines/reference/extractMethod/extractMethod1.ts
@@ -1,85 +1,98 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     let x = 1;
     function foo() {
     }
     namespace B {
         function a() {
+            let a = 1;
         
             let y = 5;
             let z = x;
-            return foo();
+            a = y;
+            foo();
         }
     }
 }
-==SCOPE::function 'a'==
+// ==SCOPE::function 'a'==
 namespace A {
     let x = 1;
     function foo() {
     }
     namespace B {
         function a() {
+            let a = 1;
         
-            return /*RENAME*/newFunction();
+            /*RENAME*/newFunction();
 
             function newFunction() {
                 let y = 5;
                 let z = x;
-                return foo();
+                a = y;
+                foo();
             }
         }
     }
 }
-==SCOPE::namespace 'B'==
+// ==SCOPE::namespace 'B'==
 namespace A {
     let x = 1;
     function foo() {
     }
     namespace B {
         function a() {
+            let a = 1;
         
-            return /*RENAME*/newFunction();
+            a = /*RENAME*/newFunction(a);
         }
 
-        function newFunction() {
+        function newFunction(a: number) {
             let y = 5;
             let z = x;
-            return foo();
+            a = y;
+            foo();
+            return a;
         }
     }
 }
-==SCOPE::namespace 'A'==
+// ==SCOPE::namespace 'A'==
 namespace A {
     let x = 1;
     function foo() {
     }
     namespace B {
         function a() {
+            let a = 1;
         
-            return /*RENAME*/newFunction();
+            a = /*RENAME*/newFunction(a);
         }
     }
 
-    function newFunction() {
+    function newFunction(a: number) {
         let y = 5;
         let z = x;
-        return foo();
+        a = y;
+        foo();
+        return a;
     }
 }
-==SCOPE::global scope==
+// ==SCOPE::global scope==
 namespace A {
     let x = 1;
     function foo() {
     }
     namespace B {
         function a() {
+            let a = 1;
         
-            return /*RENAME*/newFunction(x, foo);
+            a = /*RENAME*/newFunction(x, a, foo);
         }
     }
 }
-function newFunction(x: number, foo: () => void) {
+function newFunction(x: number, a: number, foo: () => void) {
     let y = 5;
     let z = x;
-    return foo();
+    a = y;
+    foo();
+    return a;
 }
