+++ tests/baselines/reference/extractMethod/extractMethod10.ts
@@ -1,46 +1,35 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     export interface I { x: number };
-    namespace B {
-        function a() {
+    class C {
+        a() {
+            let z = 1;
             let a1: I = { x: 1 };
             return a1.x + 10;
         }
     }
 }
-==SCOPE::function 'a'==
+// ==SCOPE::class 'C'==
 namespace A {
     export interface I { x: number };
-    namespace B {
-        function a() {
-            return /*RENAME*/newFunction();
-
-            function newFunction() {
-                let a1: I = { x: 1 };
-                return a1.x + 10;
-            }
-        }
-    }
-}
-==SCOPE::namespace 'B'==
-namespace A {
-    export interface I { x: number };
-    namespace B {
-        function a() {
-            return /*RENAME*/newFunction();
+    class C {
+        a() {
+            let z = 1;
+            return this./*RENAME*/newFunction();
         }
 
-        function newFunction() {
+        private newFunction() {
             let a1: I = { x: 1 };
             return a1.x + 10;
         }
     }
 }
-==SCOPE::namespace 'A'==
+// ==SCOPE::namespace 'A'==
 namespace A {
     export interface I { x: number };
-    namespace B {
-        function a() {
+    class C {
+        a() {
+            let z = 1;
             return /*RENAME*/newFunction();
         }
     }
@@ -50,11 +39,12 @@ namespace A {
         return a1.x + 10;
     }
 }
-==SCOPE::global scope==
+// ==SCOPE::global scope==
 namespace A {
     export interface I { x: number };
-    namespace B {
-        function a() {
+    class C {
+        a() {
+            let z = 1;
             return /*RENAME*/newFunction();
         }
     }
