+++ tests/baselines/reference/extractMethod/extractMethod11.ts
@@ -1,4 +1,4 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     let y = 1;
     class C {
@@ -11,76 +11,59 @@ namespace A {
         }
     }
 }
-==SCOPE::method a==
-namespace A {
-    let y = 1;
-    class C {
-        a() {
-            let z = 1;
-            return newFunction();
-
-            function newFunction() {
-                let a1 = { x: 1 };
-                y = 10;
-                z = 42;
-                return a1.x + 10;
-            }
-        }
-    }
-}
-==SCOPE::class C==
+// ==SCOPE::class 'C'==
 namespace A {
     let y = 1;
     class C {
         a() {
             let z = 1;
             var __return: any;
-            ({ z, __return } = this.newFunction(z));
+            ({ __return, z } = this./*RENAME*/newFunction(z));
             return __return;
         }
 
-        private newFunction(z: any) {
+        private newFunction(z: number) {
             let a1 = { x: 1 };
             y = 10;
             z = 42;
-            return { z, __return: a1.x + 10 };
+            return { __return: a1.x + 10, z };
         }
     }
 }
-==SCOPE::namespace A==
+// ==SCOPE::namespace 'A'==
 namespace A {
     let y = 1;
     class C {
         a() {
             let z = 1;
             var __return: any;
-            ({ z, __return } = newFunction(z));
+            ({ __return, z } = /*RENAME*/newFunction(z));
             return __return;
         }
     }
 
-    function newFunction(z: any) {
+    function newFunction(z: number) {
         let a1 = { x: 1 };
         y = 10;
         z = 42;
-        return { z, __return: a1.x + 10 };
+        return { __return: a1.x + 10, z };
     }
 }
-==SCOPE::file '/a.ts'==
+// ==SCOPE::global scope==
 namespace A {
     let y = 1;
     class C {
         a() {
             let z = 1;
             var __return: any;
-            ({ y, z, __return } = newFunction(y, z));
+            ({ __return, y, z } = /*RENAME*/newFunction(y, z));
             return __return;
         }
     }
 }
-function newFunction(y: any, z: any) {
+function newFunction(y: number, z: number) {
     let a1 = { x: 1 };
     y = 10;
     z = 42;
-    return { y, z, __return: a1.x + 10 };
+    return { __return: a1.x + 10, y, z };
 }
