+++ tests/baselines/reference/extractMethod/extractMethod12.ts
@@ -1,4 +1,4 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     let y = 1;
     class C {
@@ -13,7 +13,7 @@ namespace A {
         }
     }
 }
-==SCOPE::class 'C'==
+// ==SCOPE::class 'C'==
 namespace A {
     let y = 1;
     class C {
