+++ tests/baselines/reference/extractMethod/extractMethod7.ts
@@ -1,4 +1,4 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     let x = 1;
     export namespace C {
@@ -16,7 +16,7 @@ namespace A {
         }
     }
 }
-==SCOPE::function 'a'==
+// ==SCOPE::function 'a'==
 namespace A {
     let x = 1;
     export namespace C {
@@ -38,7 +38,7 @@ namespace A {
         }
     }
 }
-==SCOPE::namespace 'B'==
+// ==SCOPE::namespace 'B'==
 namespace A {
     let x = 1;
     export namespace C {
@@ -62,7 +62,7 @@ namespace A {
         }
     }
 }
-==SCOPE::namespace 'A'==
+// ==SCOPE::namespace 'A'==
 namespace A {
     let x = 1;
     export namespace C {
@@ -86,7 +86,7 @@ namespace A {
         return { __return: C.foo(), a };
     }
 }
-==SCOPE::global scope==
+// ==SCOPE::global scope==
 namespace A {
     let x = 1;
     export namespace C {
