+++ tests/baselines/reference/extractMethod/extractMethod8.ts
@@ -1,4 +1,4 @@
-==ORIGINAL==
+// ==ORIGINAL==
 namespace A {
     let x = 1;
     namespace B {
@@ -8,7 +8,7 @@ namespace A {
         }
     }
 }
-==SCOPE::function 'a'==
+// ==SCOPE::function 'a'==
 namespace A {
     let x = 1;
     namespace B {
@@ -22,7 +22,7 @@ namespace A {
         }
     }
 }
-==SCOPE::namespace 'B'==
+// ==SCOPE::namespace 'B'==
 namespace A {
     let x = 1;
     namespace B {
@@ -36,7 +36,7 @@ namespace A {
         }
     }
 }
-==SCOPE::namespace 'A'==
+// ==SCOPE::namespace 'A'==
 namespace A {
     let x = 1;
     namespace B {
@@ -50,7 +50,7 @@ namespace A {
         return 1 + a1 + x;
     }
 }
-==SCOPE::global scope==
+// ==SCOPE::global scope==
 namespace A {
     let x = 1;
     namespace B {
